--- a/src/libraries/System.Private.CoreLib/src/System/TimeZoneInfo.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/TimeZoneInfo.cs
@@ -127,6 +127,7 @@ namespace System
 
             public Dictionary<string, TimeZoneInfo>? _systemTimeZones;
             public ReadOnlyCollection<TimeZoneInfo>? _readOnlySystemTimeZones;
+            public Dictionary<string, TimeZoneInfo>? _timeZonesUsingAlternativeIds;
             public bool _allSystemTimeZonesRead;
         }
 
@@ -1898,6 +1899,9 @@ namespace System
                             }
                         }
 
+                        cachedData._timeZonesUsingAlternativeIds ??= new Dictionary<string, TimeZoneInfo>(StringComparer.OrdinalIgnoreCase);
+                        cachedData._timeZonesUsingAlternativeIds[id] = zone;
+
                         Debug.Assert(zone != null);
                         value = zone;
                     }
@@ -1926,8 +1930,25 @@ namespace System
                     }
                     else
                     {
-                        value = new TimeZoneInfo(match._id, match._baseUtcOffset, match._displayName, match._standardDisplayName,
-                                              match._daylightDisplayName, match._adjustmentRules, disableDaylightSavingTime: false, match.HasIanaId);
+                        value = match;
+                    }
+
+                    return result;
+                }
+            }
+
+            if (cachedData._timeZonesUsingAlternativeIds != null)
+            {
+                if (cachedData._timeZonesUsingAlternativeIds.TryGetValue(id, out TimeZoneInfo? match))
+                {
+                    if (dstDisabled && match._supportsDaylightSavingTime)
+                    {
+                        // we found a cache hit but we want a time zone without DST and this one has DST data
+                        value = CreateCustomTimeZone(match._id, match._baseUtcOffset, match._displayName, match._standardDisplayName);
+                    }
+                    else
+                    {
+                        value = match;
                     }
 
                     return result;
@@ -1981,8 +2002,7 @@ namespace System
                 }
                 else
                 {
-                    value = new TimeZoneInfo(match!._id, match._baseUtcOffset, match._displayName, match._standardDisplayName,
-                                          match._daylightDisplayName, match._adjustmentRules, disableDaylightSavingTime: false, match.HasIanaId);
+                    value = match;
                 }
             }
             else