Fixes https://github.com/dotnet/runtime/issues/85432

This change  is addressing multiple performance issues:

The first problem arises when creating a TimeZoneInfo object with an alternative zone ID (like using IANA Id on Windows and Windows Id on Linux) using the FindSystemTimeZoneById method. By default, the object returned by GetSystemTimeZones is cached only for the zone ID that matches the OS we are running on. Thus, when we try to create a new object with an alternative ID, we cannot benefit from the cache, and we always attempt to read the zone from the system. On non-Windows OS, this requires accessing the file system and probing for the zone file. To address this issue, we introduce a cache for the alternative ID.

The second issue concerns the unnecessary creation of new zone objects when requesting them with FindSystemTimeZoneById. Even if we have already created the object and cached it, we still create a new one every time. However, since the zone objects do not have writable properties, there is no need to clone them. Instead, we can return the cached object to avoid the allocation and the validation of the new object.

# Windows 

### Before

```
|         Method |         Mean |      Error |     StdDev |   Gen0 | Allocated |
|--------------- |-------------:|-----------:|-----------:|-------:|----------:|
|    UsingIanaId | 40,061.32 ns | 753.377 ns | 667.849 ns |      - |     312 B |
| UsingWindowsId |     50.31 ns |   1.018 ns |   1.461 ns | 0.0095 |      80 B |

```

### After 

```
|         Method |     Mean |    Error |   StdDev | Allocated |
|--------------- |---------:|---------:|---------:|----------:|
|    UsingIanaId | 34.41 ns | 0.377 ns | 0.315 ns |         - |
| UsingWindowsId | 33.55 ns | 0.462 ns | 0.410 ns |         - |
```

# Linux 

### Before

```
|         Method |       Mean |    Error |   StdDev |   Gen0 | Allocated |
|--------------- |-----------:|---------:|---------:|-------:|----------:|
|    UsingIanaId | 1,429.8 ns | 27.28 ns | 29.19 ns | 0.0095 |      80 B |
| UsingWindowsId |   935.2 ns | 18.36 ns | 21.86 ns | 0.0687 |     576 B |

```

### After

```
|         Method |     Mean |    Error |   StdDev | Allocated |
|--------------- |---------:|---------:|---------:|----------:|
|    UsingIanaId | 31.87 ns | 0.634 ns | 1.143 ns |         - |
| UsingWindowsId | 41.03 ns | 0.838 ns | 1.229 ns |         - |
```